cyber cyber cyber

+++ server.js
@@ -37,24 +37,28 @@ MongoClient.connect(url, { useNewUrlParser: true }).then(client => {
     })
 
     app.post('/answer/:conversationId', async (req, res) => {
-        try {
-            await db.collection('conversations').updateOne({ _id: ObjectId(req.params.conversationId) }, { $set: { answer: req.body.answer } })
-            console.log(`updated answer for conversation ${req.params.conversationId}`)
-            res.send()
-        } catch (err) {
-            console.log(err)
-            res.sendStatus(500)
+        if (manager) {
+            try {
+                await db.collection('conversations').updateOne({ _id: ObjectId(req.params.conversationId) }, { $set: { answer: req.body.answer } })
+                console.log(`updated answer for conversation ${req.params.conversationId}`)
+                res.send()
+            } catch (err) {
+                console.log(err)
+                res.sendStatus(500)
+            }
         }
     })
 
     app.delete('/conversation/:conversationId', async (req, res) => {
-        try {
-            await db.collection('conversations').deleteOne({ _id: ObjectId(req.params.conversationId) })
-            console.log(`deleted conversation ${req.params.conversationId} from db`)
-            res.send()
-        } catch (err) {
-            console.log(error)
-            res.sendStatus(500)
+        if (manager) {
+            try {
+                await db.collection('conversations').deleteOne({ _id: ObjectId(req.params.conversationId) })
+                console.log(`deleted conversation ${req.params.conversationId} from db`)
+                res.send()
+            } catch (err) {
+                console.log(error)
+                res.sendStatus(500)
+            }
         }
     })
 
@@ -74,6 +78,11 @@ MongoClient.connect(url, { useNewUrlParser: true }).then(client => {
             console.log(err)
             res.sendStatus(500)
         }
+    })
+
+    app.post('/disconnect', (req, res) => {
+        manager = null
+        res.send()
     })
 }).catch(console.error)
